pyproject.toml
[build-system]
requires = ["maturin>=1.11,<2.0"]
build-backend = "maturin"

[project]
name = "canvas_pyr"
description = "A Python canvas library powered by Skia, with bindings implemented in Rust."
readme = "README.md"
requires-python = ">=3.10"
dependencies = []
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
]
dynamic = ["version"]

[dependency-groups]
dev = [
    "maturin>=1.11.5,<2.0",
    "pillow==12.1.0",
    "tomli==2.4.0",
]

[tool.uv]
package = false

[tool.canvas_pyr.upstream]
name = "canvas"
version = "0.1.91"
commit = "6661e25"
url = "https://github.com/Brooooooklyn/canvas"
license = "MIT"
